sys/random/fortuna: use uint8_t for seed file
The seed file (section 9.6.1 of [1]) is defined as 64 bytes. Using
`uint32_t` made it four times too large. The correct size was already
defined as `FORTUNA_SEED_SIZE`.

This is consistent with `fortuna_write_seed` and `fortuna_update_seed`,
which never use more than `FORTUNA_SEED_SIZE`.

[1]: https://www.schneier.com/wp-content/uploads/2015/12/fortuna.pdf

+++ sys/random/fortuna/fortuna.h
@@ -179,7 +179,7 @@ typedef struct {
 /**
  * @brief Type definition of a Fortuna seed file.
  */
-typedef uint32_t fortuna_seed_t[FORTUNA_SEED_SIZE];
+typedef uint8_t fortuna_seed_t[FORTUNA_SEED_SIZE];
 
 /**
  * @brief   Initialize the Fortuna PRNG state.
